+++ jdaviz/core/events.py
@@ -7,6 +7,7 @@ __all__ = ['NewViewerMessage', 'ViewerAddedMessage', 'ViewerRemovedMessage', 'Lo
            'SliceSelectSliceMessage', 'SliceValueUpdatedMessage',
            'SliceToolStateMessage',
            'CatalogResultsChangedMessage', 'CatalogSelectClickEventMessage',
+           'FootprintSelectClickEventMessage',
            'TableClickMessage', 'LinkUpdatedMessage', 'ExitBatchLoadMessage',
            'AstrowidgetMarkersChangedMessage', 'MarkersPluginUpdate',
            'GlobalDisplayUnitChanged', 'ChangeRefDataMessage',
@@ -301,6 +302,17 @@ class CatalogSelectClickEventMessage(Message):
         self.x = x
         self.y = y
         super().__init__(*args, **kwargs)
+
+
+class FootprintSelectClickEventMessage(Message):
+    """
+    Message emitted when a user clicks on a viewer to select an overlay.
+    """
+
+    def __init__(self, data, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.x = data["domain"]["x"]
+        self.y = data["domain"]["y"]
 
 
 class RedshiftMessage(Message):