+++ CHANGES.rst
@@ -21,6 +21,9 @@ Imviz
 
 - Added ability to load remote data from a S3 URI to Imviz. [#3500]
 
+- Footprints plugin now supports selecting the closest overlay
+  to a clicked point in the image viewer. [#3525]
+
 Mosviz
 ^^^^^^